Improve error checks and messages

+++ gen.sh
@@ -6,7 +6,7 @@ check_path() {
     local path="$1"
     # echo "checking path $1"
     if [ ! -d "$path" ]; then
-        echo "ERROR: path.sconf: Path $path not valid! Exiting.."
+        echo "ERROR: path.sconf: Path $path not valid! Please run ~/ResHelper/install.sh again. Exiting.."
         exit 1
     fi
 }
@@ -16,7 +16,7 @@ SCRIPT_DIR=$(dirname "$(realpath "$0")")
 if [ -f ${SCRIPT_DIR}/paths.conf ]; then
 	. "${SCRIPT_DIR}/paths.conf"
 else
-   echo "ERROR: ${SCRIPT_DIR}/paths.conf not found. Exiting.. "
+   echo "ResHelper: ERROR: ${SCRIPT_DIR}/paths.conf not found. Please run ~/ResHelper/install.sh again. Exiting.. "
    exit 1
 fi
 
@@ -40,9 +40,17 @@ name="shaper_calibrate_$1"
 
 [ ! -d "$RES_DATA_PATH" ] && mkdir -p "$RES_DATA_PATH"
 
+# Check upfront if resonance CSV files exist
+if ls $TMP_PATH/resonances_${1}_*.csv > /dev/null 2>&1; then
+	echo "ResHelper: Found resonance .csv data for the $1 axis..."
+else
+	echo "ResHelper: ERROR: No Klipper resonance .csv data file was generated in the $TMP_PATH folder for the $1 axis!"
+	exit 1
+fi
+
 if [ "$KLIPPER_VER" = "MAIN" ] || [ "$KLIPPER_VER" = "DK" ]; then
 	if [ "$3" -eq 1 ]; then
-    	echo "Skipping Classic Mode. Required only for Kalico BE variants.\n"
+    	echo "ResHelper: Skipping Classic Mode. Required only for Kalico BE variants.\n"
 	fi
     echo "ResHelper: Starting Klipper Graph Generation...\n"
     ${PK_PATH}/python ${KLIPPER_PATH}/scripts/calibrate_shaper.py "${TMP_PATH}"/resonances_"$1"_*.csv -o "${RES_DATA_PATH}"/shaper_calibrate_"$1".png --shapers zv,mzv,ei
@@ -54,9 +62,17 @@ elif [ "$KLIPPER_VER" = "DK_BE" ]; then
 	    echo "ResHelper Kalico BE: Starting Klipper Graph Generation...\n"
 	    ${PK_PATH}/python ${KLIPPER_PATH}/scripts/calibrate_shaper.py "${TMP_PATH}"/resonances_"$1"_*.csv -o "${RES_DATA_PATH}"/shaper_calibrate_"$1".png
 	elif [ "$3" -eq 1 ]; then
-	    # Classic klipper generation
-	    echo "ResHelper Kalico BE: Starting Classic Klipper Graph Generation...\n"
-	    ${PK_PATH}/python ${KLIPPER_PATH}/scripts/calibrate_shaper_classic.py "${TMP_PATH}"/resonances_"$1"_*.csv -o "${RES_DATA_PATH}"/shaper_calibrate_"$1".png --shapers zv,mzv,ei --classic true
+			# Classic klipper generation
+			echo "ResHelper Kalico BE: Starting Classic Klipper Graph Generation...\n"
+
+			if [ -f "${KLIPPER_PATH}/scripts/calibrate_shaper_classic.py" ]; then
+					${PK_PATH}/python ${KLIPPER_PATH}/scripts/calibrate_shaper_classic.py \
+							"${TMP_PATH}"/resonances_"$1"_*.csv \
+							-o "${RES_DATA_PATH}"/shaper_calibrate_"$1".png \
+							--shapers zv,mzv,ei --classic true
+			else
+					echo "Reshelper: Error: no /scripts/calibrate_shaper_classic.py module found in Kalico. Please run ~/ResHelper/install.sh again!"
+			fi
 	else
 	    # Handle unexpected values of $3
 	    echo "Invalid value for third 'Classic' argument: $3. Expected 0 or 1."
@@ -64,39 +80,47 @@ elif [ "$KLIPPER_VER" = "DK_BE" ]; then
 	fi
 
 else
-	echo "ERROR: $KLIPPER_VER - Could not determine Klipper variant. Exiting.."
+	echo "ERROR: $KLIPPER_VER - Could not determine Klipper variant. Please run ~/ResHelper/install.sh again. Exiting.."
 	exit 1
 fi
 
 # Damping ratio
 if [ "$2" -eq 1 ]; then 
-    echo "ResHelper : Calculating damping ratio for $1"
+		echo "ResHelper: Calculating damping ratio for the $1 axis"
 
-	PYTHON=$(command -v python3 || command -v python)
-	if [ -z "$PYTHON" ]; then
-	    echo "Neither python nor python3 is installed."
-	    exit 1
-	fi
-	
-	DR_RESULT=$($PYTHON "${RH_PATH}/dr_solver.py" "${TMP_PATH}/resonances_${1}_*.csv")
+				PYTHON=$(command -v python3 || command -v python)
+				if [ -z "$PYTHON" ]; then
+						echo "ResHelper: Neither python nor python3 is installed."
+				else
+						DR_RESULT=$($PYTHON "${RH_PATH}/dr_solver.py" "$TMP_PATH/resonances_${1}_*.csv")
 
-	# Validating results
-	if [ "$(echo "$DR_RESULT > 0.001" | bc -l)" -eq 1 ] && [ "$(echo "$DR_RESULT < 2.0" | bc -l)" -eq 1 ]; then
-	    dr=$DR_RESULT
-	    echo "ResHelper : Damping ratio calculated:\ndamping_ratio: $dr\n"
-	else
-	    echo "WARNING: The DR result is out of expected range: $DR_RESULT \n"
-	fi
-    
-fi
+						# Validating results
+						if [ "$(echo "$DR_RESULT > 0.001" | bc -l)" -eq 1 ] && \
+							[ "$(echo "$DR_RESULT < 2.0" | bc -l)" -eq 1 ]; then
+								dr=$DR_RESULT
+								echo "ResHelper: Damping ratio calculated: $DR_RESULT"
+						else
+								echo "ResHelper: WARNING: The DR result is out of expected range: $DR_RESULT"
+						fi
+				fi
+		fi
+
 
 # Cleanup
-name="$name-dr_${dr:-NA}-v$(date "+%Y%m%d_%H%M").png"
-mv "$RES_DATA_PATH"/shaper_calibrate_"$1".png "$RES_DATA_PATH/$name"
+if [ -f "$RES_DATA_PATH/shaper_calibrate_$1.png" ]; then
+		name="$name-dr_${dr:-NA}-v$(date "+%Y%m%d_%H%M").png"
+    mv "$RES_DATA_PATH/shaper_calibrate_$1.png" "$RES_DATA_PATH/$name"
+    echo "ResHelper: Image generated: $RES_DATA_PATH/$name"
+else
+    echo "ResHelper: ERROR: Klipper's calibrate_shaper script failed to generate an image graph!"
+fi
 
-rm "${TMP_PATH}"/rh-prev-run/*.csv 2>/dev/null
-mkdir -p ${TMP_PATH}/rh-prev-run
+if ls $TMP_PATH/resonances_*.csv > /dev/null 2>&1; then
+	rm "${TMP_PATH}"/rh-prev-run/*.csv 2>/dev/null
+	mkdir -p ${TMP_PATH}/rh-prev-run
+	cp "$TMP_PATH"/resonances_*.csv "${TMP_PATH}/rh-prev-run/" && rm "$TMP_PATH"/resonances_*.csv
+else
+	echo "ResHelper: No /tmp file cleanup needed"
+fi
 
-cp "$TMP_PATH"/resonances_*.csv "${TMP_PATH}/rh-prev-run/" && rm "$TMP_PATH"/resonances_*.csv
-
-echo "ResHelper: Finished\n"
+echo "ResHelper: Finished"
